Split set_error out of Metadata::Tagging
